pyramid: docs and protected imports
